requirements.txt
# AOMaaS - Autonomous Open-Source Maintainer as a Service
# Core Dependencies

# Web Framework & API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
starlette>=0.40.0

# Data Validation & Models
pydantic>=2.5.0
pydantic-settings>=2.0.0
annotated-types>=0.6.0

# Async & HTTP
httpx>=0.25.0
aiofiles>=23.2.0
anyio>=4.0.0

# Database & Storage
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0

# Cache & Message Queue
redis>=5.0.0
celery>=5.3.0
kombu>=5.5.0
billiard>=4.2.0

# Vector Database
qdrant-client>=1.7.0

# Object Storage
minio>=7.2.0

# GitHub Integration
pygithub>=2.1.0
gitpython>=3.1.40

# AI Services
openai>=1.3.0
anthropic>=0.7.0
tiktoken>=0.5.0

# Code Analysis
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-typescript>=0.20.0
tree-sitter-rust>=0.20.0
tree-sitter-go>=0.20.0
tree-sitter-java>=0.20.0

# CLI & UI
typer>=0.9.0
rich>=13.7.0
click>=8.1.0
shellingham>=1.3.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Logging & Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Template Engine
jinja2>=3.1.0

# Utilities
python-dateutil>=2.8.0
requests>=2.31.0
charset-normalizer>=3.0.0
urllib3>=2.0.0
certifi>=2023.0.0
idna>=3.4
sniffio>=1.3.0

# Security
cryptography>=41.0.0
argon2-cffi>=23.1.0

# Development Dependencies (Optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0
ruff>=0.1.0
pre-commit>=3.6.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0
markdown-it-py>=3.0.0
mdurl>=0.1.0
pygments>=2.16.0

# Testing & Quality
coverage>=7.3.0
factory-boy>=3.3.0
faker>=20.0.0

# Async Testing (FIXED)
pytest-httpx>=0.35.0
respx>=0.20.0

# Time & Date
tzdata>=2023.3

# Workflow & Automation
pydantic-yaml>=1.2.0
jsonschema>=4.19.0

# Performance
orjson>=3.9.0
msgpack>=1.0.0

# Optional: ML/AI Dependencies
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0

# Development Tools
ipython>=8.15.0
jupyter>=1.0.0

# Container Health Checks  
psutil>=5.9.0

# Git Operations
dulwich>=0.21.0

# File Watching (for development)
watchfiles>=0.21.0
